requirements.txt
discord.py
gTTS
PyNaCl
arrow